% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_input_format}
\alias{check_input_format}
\title{Input format control on uploading files}
\usage{
check_input_format(inputdf, type)
}
\arguments{
\item{inputdf}{A data.table object for RNA-seq SVs, DNA-seq SVs or mutation profile.}

\item{type}{A character string (e.g. 'RNA', 'DNA' or 'Mut').}
}
\description{
Input format QC of upload files.
}
