% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_persample_bam.R
\name{chrom_name_function}
\alias{chrom_name_function}
\title{Convert chromome notation (e.g. '1' to 'chr1')}
\usage{
chrom_name_function(file, selection)
}
\arguments{
\item{file}{path of read alignment (BAM format).}

\item{selection}{A GRanges object defines a region for coverage plotting using the bamfile.}
}
\value{
A GRanges object with read alignment in the selection region.
}
\description{
This function it to convert chromosome notation in import bam file, which meets requirement of Gviz::AlignmentsTrack function
}
