% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuSViz.R
\docType{data}
\name{configure}
\alias{configure}
\title{FuSViz}
\format{
An object of class \code{NULL} of length 0.
}
\usage{
configure
}
\description{
Make a HTML widget for an embeddable interactive genome visualization javascript library (igv.js)
}
\keyword{datasets}
