% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuSViz.R
\name{Coordinate}
\alias{Coordinate}
\title{Genomic coordinate retrieve}
\usage{
Coordinate(session, inputid)
}
\arguments{
\item{session}{An environment object that can be used to access information and functionality by shiny.}

\item{inputid}{A string represents the name of DOM element (inputId) defined in ui.R (e.g. \code{textInput(inputId, label, value = "")}).}
}
\description{
Obtain genomic coordinate of a zoom region
}
