% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_persample.R
\name{.plot_curve}
\alias{.plot_curve}
\title{Plot curved lines to link breakpoints}
\usage{
.plot_curve(
  upstream_xy,
  upstream_flag,
  downstream_xy,
  downstream_flag,
  breakpoint_xy
)
}
\arguments{
\item{upstream_xy}{A data.frame (i.e. a returned object \code{'A1_xy'} in FUNCTION "plot_separate_individual" - exon coordinates(x1, y1, x2, y2) of geneA transcripts).}

\item{upstream_flag}{A list (i.e. the same object \code{'first'} used in FUNCTION "plot_separate_individual"; \code{'first\$pos\$transcript'} is a one-row data.frame).}

\item{downstream_xy}{A data.frame (i.e. a returned object \code{'B1_xy'} in FUNCTION "plot_separate_individual" - exon coordinates(x1, y1, x2, y2) of geneB transcripts).}

\item{downstream_flag}{A list (i.e. the same object \code{'second'} used in FUNCTION "plot_separate_individual"; \code{'second\$pos\$transcript'} is a one-row data.frame).}

\item{breakpoint_xy}{A data.frame with one row record (i.e. \code{'breakpoint_xy[i,1]'} - breakpoint pos of geneA; \code{'breakpoint_xy[i,2]'} - breakpoint pos of geneB).}
}
\value{
A data.frame with three columns (i.e. \code{'x_pos_gene_upstream'}, \code{'x_pos_gene_downstream'}, \code{'y_pos'})
}
\description{
Plot curved lines to link breakpoints of geneA and geneB transcripts in 'per_sample' tab-panel
}
