% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_overview.R
\name{.plot_curve_overview}
\alias{.plot_curve_overview}
\title{Plot curved lines to link breakpoints}
\usage{
.plot_curve_overview(
  upstream_xy,
  upstream_flag,
  downstream_xy,
  downstream_flag,
  breakpoint_xy
)
}
\arguments{
\item{upstream_xy}{A data.frame (i.e. a returned object \code{'collect\$A1_xy'} in FUNCTION "plot_arrow_overview" - exon coordinates(x1, y1, x2, y2) of geneA transcripts).}

\item{upstream_flag}{A list (i.e. the same object \code{'first'} used in FUNCTION "plot_arrow_overview"; \code{'first\$pos\$transcript'} is a one-row data.frame).}

\item{downstream_xy}{A data.frame (i.e. a returned object \code{'collect\$B1_xy'} in FUNCTION "plot_arrow_overview" - exon coordinates(x1, y1, x2, y2) of geneB transcripts).}

\item{downstream_flag}{A list (i.e. the same object \code{'second'} used in FUNCTION "plot_arrow_overview"; \code{'second\$pos\$transcript'} is a one-row data.frame).}

\item{breakpoint_xy}{A data.frame with one row (i.e. 'breakpoint_xy\link{i,1}' - breakpoint pos in geneA; 'breakpoint_xy\link{i,2}' - breakpoint pos in geneB;
'breakpoint_xy\link{i,3}' is the number of samples with such fusion).}
}
\value{
A data.frame with three columns (i.e. \code{'x_pos_gene_upstream'}, \code{'x_pos_gene_downstream'}, \code{'y_pos'})
}
\description{
Plot curved lines to link breakpoints of geneA and geneB transcripts in 'overview' tab-panel
}
