% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuSViz.R
\name{DuplicateTrackRemove}
\alias{DuplicateTrackRemove}
\title{Remove track with replicate names}
\usage{
DuplicateTrackRemove(session, name)
}
\arguments{
\item{session}{An environment object that is used to access information and functionality by shiny.}

\item{name}{A character vector represents track name.}
}
\description{
Remove track with replicate names
}
