% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuSViz.R
\name{FuSViz}
\alias{FuSViz}
\title{FuSViz widget}
\usage{
FuSViz(
  genomeName,
  trackHeight = 300,
  displayMode = "EXPANDED",
  initialLocus = NULL,
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{genomeName}{A string shows human genome reference version, either \code{hg19} or \code{hg38}.}

\item{trackHeight}{A number defines initial height of a track (e.g. default value: \code{300}).}

\item{displayMode}{Annotation track display mode (e.g. \code{COLLAPSED}, \code{EXPANDED}, \code{SQUISHED}).}

\item{initialLocus}{A string defines initial genomic location when widget loading is launched (e.g. default value: \code{TP53}).}

\item{width, height}{Must be a valid CSS unit (like \code{100\%},
\code{400px}, \code{auto}) or a number, which will be coerced to a string and have \code{px} appended.}

\item{elementId}{A string represents the name of a HTML element id used for DOM manipulation.}
}
\description{
Defined a main class to build widget
}
