% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuSViz.R
\name{FuSVizOutput}
\alias{FuSVizOutput}
\title{Shiny bindings for FuSViz}
\usage{
FuSVizOutput(outputId, width = "100\%", height = "400px")
}
\arguments{
\item{outputId}{Manipulate as a DOM element id}

\item{width, height}{Must be a valid CSS unit (like \code{100\%},
\code{400px}, \code{auto}) or a number, which will be coerced to a string and have \code{px} appended.}
}
\description{
Output element embedded in ui { }
}
