% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{FuVSiz_app}
\alias{FuVSiz_app}
\title{Start FuSViz web app}
\usage{
FuVSiz_app(data = NULL)
}
\arguments{
\item{data}{Data to be loaded with app; the default value as NULL}
}
\description{
The app is started with system default browser
}
