% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Breakpoint_persample.R
\name{gene_trans_ex}
\alias{gene_trans_ex}
\title{Tag breakpoint position in each transcript of partner genes}
\usage{
gene_trans_ex(breakpoint, object, whole_tx)
}
\arguments{
\item{breakpoint}{A numeric vector (containing only one element) represents genomic coordinate of geneA / geneB breakpoint.}

\item{object}{A list collects geneA / geneB breakpoint annotation, two keys elements (i.e. a GRange object: \code{object\$txTr_f}, a data.frame object: \code{object\$dataset}).}

\item{whole_tx}{A GRange object - exon intervals grouped by transcript per gene, which is generated by processing \code{txdb_ref} object.}
}
\value{
A list with two data.frame elements (i.e. \code{transcript} and \code{select_region}) with transcript and breakpoint annotations of partner genes
}
\description{
Tag breakpoint position in each transcript (e.g. at exon boundary/within exon or within intron, exon rank) of partner gene
for two-way plotting in per_sample tab-panel
}
