% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Annotation.R
\name{get_annotation_db}
\alias{get_annotation_db}
\title{Load gene-model annotations compiled from GTF format}
\usage{
get_annotation_db(ensemblId, txdb_ref, grTrack_ref)
}
\arguments{
\item{ensemblId}{A string - ensembl gene id (e.g. \emph{ENSG00000141510}).}

\item{txdb_ref}{An object - an annotation database engine loaded from *.sqlite file (initially created by compiling a GTF annotation file).}

\item{grTrack_ref}{An object - annotation tracks loaded from *.Rd file (originally generated by processing \code{txdb_ref} object).}
}
\value{
A list (e.g. object_A and object_B) to collect transcript annotations of partner genes
}
\description{
Load gene-model annotations for two-way plotting in per_sample and domain tab-panels
}
