% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Network_process.R
\name{network_process}
\alias{network_process}
\title{Create the input format for visNetwork package}
\usage{
network_process(
  tmp,
  type,
  color_onco,
  color_supp,
  color_rela,
  color_inter,
  color_other
)
}
\arguments{
\item{tmp}{A data.frame with three columns (i.e. \code{'gene1'}, \code{'gene2'}, \code{'name'}).}

\item{type}{A string (e.g. \code{'DNA'} or \code{'RNA'}).}

\item{color_onco}{A string of color value (e.g. \code{'#ff0000'} or \code{'rgb(100,0,100)'}); if partner gene is oncogene (defalut: '#ff8566').}

\item{color_supp}{A string of color value (e.g. \code{'#ff0000'} or \code{'rgb(100,0,100)'}); if partner gene is tumor suppress gene (defalut: '#00ccff').}

\item{color_rela}{A string of color value (e.g. \code{'#ff0000'} or \code{'rgb(100,0,100)'}); if partner gene is cancer-related gene (defalut: '#ffcc33').}

\item{color_inter}{A string of color value (e.g. \code{'#ff0000'} or \code{'rgb(100,0,100)'}); if breakpoint is in intergenic region (defalut: '#262626').}

\item{color_other}{A string of color value (e.g. \code{'#ff0000'} or \code{'rgb(100,0,100)'}); if not belonging to any category above (defalut: '#f2f2f2').}
}
\value{
A list with three data.frame elements (i.e. \code{'nodes'}, \code{'edges'} and \code{'degree_score'}).
}
\description{
Create the input format of DNA and RNA SVs for visNetwork visualization
}
