% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_domain.R
\name{plot_separate_domain_geneA}
\alias{plot_separate_domain_geneA}
\title{Plot transcript-model(reducing introns), domain and motif}
\usage{
plot_separate_domain_geneA(first, first_name, first_domain, first_motif)
}
\arguments{
\item{first}{A list: \strong{'key'} is a string \code{'\$pos'} (i.e. breakpoint pos in geneA); \code{'first\$pos\$transcript'} is a data.frame object
that collects evaluation of breakpoint pos in geneA transcripts after reducing intron; \code{'first\$pos\$select_region'} is a data.frame object
(5utr-cds-utr3 annotation after reducing intron) for plotting and it is a constant variable for different \code{'\$pos'} values.}

\item{first_name}{A string - symbol name of geneA (e.g. 'TMPRSS2').}

\item{first_domain}{A data.frame for collecting domain annotation (e.g. \code{'first_domain\$Start'}: domain start; \code{'first_domain\$End'}: domain end;
\code{'first_domain\$Domain_name'}: domain name abbreviation).}

\item{first_motif}{A data.frame for collecting motif annotation (e.g. \code{'first_motif\$Start'}: motif start; \code{'first_motif\$End'}: motif end;
\code{'first_motif\$Domain_name'}: motif name abbreviation).}
}
\value{
A data.frame (i.e. \code{'A1_xy'}) that contains exon coordinates in the plot layout.
}
\description{
Plot transcript-model(reducing introns), domain and motif of geneA in 'domain' tab-panel
}
