% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_domain.R
\name{plot_separate_domain_geneB}
\alias{plot_separate_domain_geneB}
\title{Plot transcript-model(reducing introns), domain and motif}
\usage{
plot_separate_domain_geneB(
  second,
  second_name,
  second_domain,
  second_motif,
  offset = 0
)
}
\arguments{
\item{second}{A list - \strong{'key'} is \code{'\$pos'} (i.e. breakpoint pos of geneB); \code{'second\$pos\$transcript'} is a data.frame object
that collects evaluation of breakpoint pos for geneB transcript after reducing intron; \code{'second\$pos\$select_region'} is a data.frame object
(5utr-cds-utr3 annotation after reducing intron) for plotting and it is a constant variable for different \code{'\$pos'} values.}

\item{second_name}{A string - symbol name of geneB (e.g. 'ERG').}

\item{second_domain}{A data.frame for collecting domain annotation (e.g. \code{'second_domain\$Start'}: domain start; \code{'second_domain\$End'}: domain end;
\code{'second_domain\$Domain_name'}: domain name abbreviation).}

\item{second_motif}{A data.frame for collecting motif annotation (e.g. \code{'second_motif\$Start'}: motif start; \code{'secondt_motif\$End'}: motif end;
\code{'second_motif\$Domain_name'}: motif name abbreviation).}

\item{offset}{A numeric value - extend the partner gene interval (default: 0, e.g. \link{start-offset, end+offset}).}
}
\value{
A data.frame (i.e. \code{'B1_xy'}) that contains exon coordinates in the plot layout.
}
\description{
Plot transcript-model(reducing introns), domain and motif of geneB in 'domain' tab-panel
}
