% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_persample_bam.R
\name{plot_separate_individual_bam}
\alias{plot_separate_individual_bam}
\title{Plot fusion transcript in context of chromosome ideogram, gene annotation and read coverage}
\usage{
plot_separate_individual_bam(
  first_name,
  second_name,
  breakpoint_A,
  breakpoint_B,
  rna_bam_path,
  dna_bam_path = NULL,
  version = "hg38",
  coverage_plot_trans = F,
  offset = 2000,
  chrom_notation_rna = T,
  chrom_notation_dna = T,
  duplicate = F,
  unmappedmate = NA,
  notpassQC = F,
  properpair = NA,
  split = NULL,
  span = NULL,
  transcriptA = NULL,
  transcriptB = NULL,
  fusion_strandA = NULL,
  fusion_strandB = NULL,
  coverage_max_A = NULL,
  coverage_max_B = NULL
)
}
\arguments{
\item{first_name}{A string - symbol name of geneA (e.g. 'TMPRSS2').}

\item{second_name}{A string - symbol name of geneB (e.g. 'ERG').}

\item{breakpoint_A}{A number - genomic coordinate of breakpoint in geneA.}

\item{breakpoint_B}{A number - genomic coordinate of breakpoint in geneB.}

\item{rna_bam_path}{A string - path of alignment file from RNA-seq (BAM format).}

\item{dna_bam_path}{A string - path of alignment file from DNA-seq (BAM format), default value: NULL.}

\item{version}{A string - genomic annotation version (either 'hg38' or 'hg19', default: 'hg38')}

\item{coverage_plot_trans}{A boolean type - if TRUE, RNA read coverage is plotted by reads mapped to exons and related introns of geneA and geneB transcripts; otherwise (default: FALSE) RNA read coverage is plotted by reads mapped to full genomic regions of geneA and geneB. coverage_plot_trans is only valid for RNA-seq alignment.}

\item{offset}{A number - the value (default: 2000) indicates how many bases are extended from 5'-end and 3'-end of geneA and geneB in plotting.}

\item{chrom_notation_rna}{A boolean type - if TRUE (default value) chromosome name in RNA-seq alignment denotes like 'chrX'; if FALSE chromosome denotes like 'X'.}

\item{chrom_notation_dna}{A boolean type - if TRUE (default value) chromosome name in DNA-seq alignment denotes like 'chrX'; if FALSE chromosome denotes like 'X'.}

\item{duplicate}{A boolean type - a logical value indicating that un-duplicated (FALSE), duplicated (TRUE), or any (NA) reads should be returned (default: F).}

\item{unmappedmate}{A boolean type - a logical value indicating whether reads with mapped (FALSE), unmapped (TRUE), or any (NA) mate should be returned (default: NA).}

\item{notpassQC}{A boolean type - a logical value indicating whether reads passing QC (FALSE), reads not passing QC (TRUE), or any (NA) reads should be returned (default: F).}

\item{properpair}{A boolean type - a logical value indicating whether improperly paired (FALSE), properly paired (TRUE), or any (NA) reads should be returned (default: NA).}

\item{split}{A number - split reads support (default: NULL).}

\item{span}{A number - discordant spanning read pair support (default: NULL).}

\item{transcriptA}{A string - selected transcripts (ensembl id) of geneA for plotting (default: NULL; "id1 id2 id3" if multiple ensembl id available).}

\item{transcriptB}{A string - selected transcripts (ensembl id) of geneB for plotting (default: NULL; "id1 id2 id3" if multiple ensembl id available).}

\item{fusion_strandA}{A string - the strand direction of geneA in transcribed fusion (default: NULL; options: '+' or '-').}

\item{fusion_strandB}{A string - the strand direction of geneB in transcribed fusion (default: NULL; options: '+' or '-').}

\item{coverage_max_A}{A number - the max value of y-axis (RNA-seq read coverage) scale for geneA (default: NULL).}

\item{coverage_max_B}{A number - the max value of y-axis (RNA-seq read coverage) scale for geneB (default: NULL).}
}
\description{
Plot curved lines to link breakpoints of fusion transcripts in context of gene annotation and read coverage
}
