% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_overview.R
\name{plot_separate_overview}
\alias{plot_separate_overview}
\title{Plot transcript-model and chromosome ideogram of partner genes}
\usage{
plot_separate_overview(
  first,
  first_name,
  second,
  second_name,
  cytoband,
  offset = 2000
)
}
\arguments{
\item{first}{A list - \strong{'key'} is \code{'\$pos'} (i.e. breakpoint pos of geneA); \code{'first\$pos\$transcript'} is a data.frame object
(only one row) that collects evaluation of breakpoint pos for geneA transcripts; \code{'first\$pos\$select_region'} is a data.frame object
(5utr-cds-utr3 annotation) for plotting and it is a constant variable for different \code{'\$pos'} values.}

\item{first_name}{A string - symbol name of geneA (e.g. 'TMPRSS2').}

\item{second}{A list - \strong{'key'} is \code{'\$pos'} (i.e. breakpoint pos of geneB); \code{'second\$pos\$transcript'} is a data.frame object
(only one row) that collects evaluation of breakpoint pos for geneB transcripts; \code{'second\$pos\$select_region'} is a data.frame object
(5utr-cds-utr3 annotation) for plotting and it is a constant variable for different \code{'\$pos'} values.}

\item{second_name}{A string - symbol name of geneB (e.g. 'ERG').}

\item{cytoband}{A GRange object for plotting cytoband of chromosome ideogram.}

\item{offset}{A numeric value - extend the partner gene interval (default: 2000, e.g. \link{start-offset, end+offset}).}
}
\value{
A list with two data.frame elements (i.e. \code{'A1_xy'} and \code{'B1_xy'}) that contain exon coordinates in the plot layout.
}
\description{
Plot transcript-model and chromosome ideogram of partner genes in 'overview' tab-panel
}
