% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_overview.R
\name{plot_separate_overview_download}
\alias{plot_separate_overview_download}
\title{Plot transcript-model, chromosome ideogram and multiple curved lines that link breakpoints of geneA and geneB transcripts (plot downloading version)}
\usage{
plot_separate_overview_download(
  first,
  first_name,
  second,
  second_name,
  cytoband,
  breakpoint
)
}
\description{
Plot transcript-model, chromosome ideogram and multiple curved lines that link breakpoints of geneA and geneB transcripts together in 'overview' tab-panel
}
