% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuSViz.R
\name{renderFuSViz}
\alias{renderFuSViz}
\title{Shiny bindings for FuSViz}
\usage{
renderFuSViz(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{An expression that generates a FuSViz.}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{FALSE} or \code{TRUE})? This
is useful if you want to save an expression in a variable.}
}
\description{
Render igv browser element in server { }
}
