% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuSViz.R
\name{TrackinBAM}
\alias{TrackinBAM}
\title{Load a BAM format from url}
\usage{
TrackinBAM(session, bam, bamindex, trackHeight = 200)
}
\arguments{
\item{session}{An environment object that can be used to access information and functionality by shiny.}

\item{bam}{An url link of BAM file}

\item{bamindex}{An url link of BAM index file}

\item{trackHeight}{Initial height of track viewport in pixels (e.g. \code{200}).}
}
\description{
Load a BAM format from url
}
