% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuSViz.R
\name{TrackinBed}
\alias{TrackinBed}
\title{Load a bed format track}
\usage{
TrackinBed(
  session,
  datastr,
  name,
  color = "gray",
  trackHeight = 100,
  displayMode = "EXPANDED",
  SameNameDel = TRUE
)
}
\arguments{
\item{session}{An environment object that can be used to access information and functionality by shiny.}

\item{datastr}{A data.frame with at least three columns (e.g. \code{chr}, \code{start}, \code{end}, from the fourth to the last column is not essential).}

\item{name}{A string of track name (e.g. \code{RNA breakpoint track}).}

\item{color}{CSS color value for track feature (e.g. \code{#ff0000} or \code{rgb(100,0,100)}).}

\item{trackHeight}{Initial height of track viewport in pixels (e.g. \code{100}).}

\item{displayMode}{Annotation track display mode (e.g. \code{COLLAPSED}, \code{EXPANDED}, \code{SQUISHED}).}

\item{SameNameDel}{A logical value controls whether to add a track if its name has been existed (e.g. \code{FALSE} or \code{TRUE}).}
}
\description{
Load a bed format track representing SV breakpoint interval
}
