% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuSViz.R
\name{TrackinBedGraph}
\alias{TrackinBedGraph}
\title{Load a bedgraph format track}
\usage{
TrackinBedGraph(
  session,
  datastr,
  name,
  color = "gray",
  trackHeight = 50,
  autoscale = TRUE,
  min = 0,
  max = 50,
  displayMode = "EXPANDED",
  SameNameDel = TRUE
)
}
\arguments{
\item{session}{An environment object that is used to access information and functionality by shiny.}

\item{datastr}{A data.frame with four columns (e.g. \code{chr}, \code{start} and \code{value}).}

\item{name}{A string - track name (e.g. \code{RNA breakpoint freq track}).}

\item{color}{CSS color value for track feature (e.g. \code{#ff0000} or \code{rgb(100,0,100)}).}

\item{trackHeight}{Initial height of track viewport in pixels (e.g. \code{100}).}

\item{autoscale}{A logical value of y-axis scale (e.g. \code{FALSE} or \code{TRUE}).}

\item{min}{A number represents minimum value of y-axis.}

\item{max}{A number represents maximum value of y-axis.}

\item{displayMode}{Display mode of annotation track (e.g. \code{COLLAPSED}, \code{EXPANDED}, \code{SQUISHED}).}

\item{SameNameDel}{A logical value controls whether to add track if its name has been existed (e.g. \code{FALSE} or \code{TRUE}).}
}
\description{
Load a bedgraph format track - count the freq of SV breakpoints
}
