% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuSViz.R
\name{TrackinBedPe}
\alias{TrackinBedPe}
\title{Load a bedpe format track}
\usage{
TrackinBedPe(
  session,
  datastr,
  name,
  color = "blue",
  thickness = 0.5,
  trackHeight = 200,
  logScale = TRUE,
  alpha = 0,
  min = 0,
  displayMode = "EXPANDED",
  SameNameDel = TRUE
)
}
\arguments{
\item{session}{An environment object that is used to access information and functionality by shiny.}

\item{datastr}{A data.frame with nine columns, e.g. \code{chr1}, \code{start1}, \code{end1}, \code{chr2}, \code{start2}, \code{end2},
\code{name}(multiple samples separated by ";"), \code{score}(number of samples), \code{type}.}

\item{name}{A string - track name (e.g. \code{DNA bedpe track}).}

\item{color}{CSS color value for track feature (e.g. \code{#ff0000} or \code{rgb(100,0,100)}).}

\item{thickness}{A number represents thickness of curve (e.g. \code{0.5}).}

\item{trackHeight}{Initial height of track viewport in pixels (e.g. \code{100}).}

\item{logScale}{A logical value (e.g. \code{FALSE} or \code{TRUE}).}

\item{alpha}{A number (0-1) that controls arc of curve (e.g. \code{0}).}

\item{min}{A number represents minimum value of y-axis.}

\item{displayMode}{Display mode of annotation track (e.g. \code{COLLAPSED}, \code{EXPANDED}, \code{SQUISHED}).}

\item{SameNameDel}{A logical value controls whether to add track if its name has been existed (e.g. \code{FALSE} or \code{TRUE}).}
}
\description{
Load a bedpe format track - count the distribution of SVs (type including: DEL, DUP, INS and INV; BND is not shown due to translocation)
}
