% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuSViz.R
\name{TrackinFile}
\alias{TrackinFile}
\title{Load local userdefined file}
\usage{
TrackinFile(session, fileobj, trackHeight = 100)
}
\arguments{
\item{session}{An environment object that can be used to access information and functionality by shiny.}

\item{fileobj}{A string shows a raw path of upload file in shiny.}

\item{trackHeight}{Initial height of track viewport in pixels (e.g. \code{100}).}
}
\description{
Load local userdefined file (only three types - vcf, bed and gtf are supported)
}
