% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuSViz.R
\name{TrackinSeg}
\alias{TrackinSeg}
\title{Load a seg format track}
\usage{
TrackinSeg(
  session,
  datastr,
  name,
  trackHeight = 50,
  color = "red",
  isLog = TRUE,
  displayMode = "EXPANDED",
  SameNameDel = TRUE
)
}
\arguments{
\item{session}{An environment object that can be used to access information and functionality by shiny.}

\item{datastr}{A data.frame with five columns (e.g. \code{chr}, \code{start}, \code{end}, \code{value} and \code{sample}).}

\item{name}{A string of track name (e.g. \code{DNA seg track}).}

\item{trackHeight}{Initial height of track viewport in pixels (e.g. \code{100}).}

\item{color}{CSS color value for track features (e.g. \code{#ff0000} or \code{rgb(100,0,100)}).}

\item{isLog}{A logical value (e.g. \code{FALSE} or \code{TRUE} - if segment values are 2*log2(copyNumber/2)).}

\item{displayMode}{Annotation display mode (e.g. \code{COLLAPSED}, \code{EXPANDED}, \code{SQUISHED}).}

\item{SameNameDel}{A logical value controls whether to add track if its name has been existed (e.g. \code{FALSE} or \code{TRUE}).}
}
\description{
Load a seg format track - count the DEL and DUP events of DNA SVs (small copy number variations)
}
