% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuSViz.R
\name{Trackoffline}
\alias{Trackoffline}
\title{Load gene track in offline mode}
\usage{
Trackoffline(
  session,
  version,
  name,
  trackHeight = 200,
  displayMode = "EXPANDED",
  SameNameDel = TRUE
)
}
\arguments{
\item{session}{An environment object that is used to access information and functionality by shiny.}

\item{version}{A string shows genome reference version (either 'hg19' or 'hg38')}

\item{name}{Initial height of track viewport in pixels (e.g. \code{100}).}

\item{trackHeight}{Initial height of track viewport in pixels (e.g. \code{100}).}

\item{displayMode}{Display mode of annotation track (e.g. \code{COLLAPSED}, \code{EXPANDED}, \code{SQUISHED}).}

\item{SameNameDel}{A logical value controls whether to add a track if its name has been existed (e.g. \code{FALSE} or \code{TRUE}).}
}
\description{
Load gene track in offline mode (only for hg19_offline or hg38_offline)
}
