% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuSViz.R
\name{UserTracksRemove}
\alias{UserTracksRemove}
\title{Remove track added by users}
\usage{
UserTracksRemove(session)
}
\arguments{
\item{session}{An environment object that is used to access information and functionality by shiny.}
}
\description{
Remove track added by users
}
