% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_input_format}
\alias{check_input_format}
\title{Input format control on uploading files}
\usage{
check_input_format(inputdf, type)
}
\description{
Input format control on uploading file (e.g. RNA SVs, DNA SVs, mutations)
}
