% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hist_process.R
\name{histgram_process}
\alias{histgram_process}
\title{Divide sample size to four parts}
\usage{
histgram_process(hist)
}
\arguments{
\item{hist}{A numeric vector contains SV number counted per sample (named by sample id).}
}
\value{
A list with four elements (i.e. \code{'section1'}, \code{'section2'}, \code{'section3'} and \code{'section4'} - each represents a SV distribution in one-fourth of samples)
}
\description{
Divide sample size to four equal parts for histgram plotting the SV distribution
}
