% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_domain.R
\name{plot_separate_domain_download}
\alias{plot_separate_domain_download}
\title{Plot transcript-model(reducing introns), domain, motif and multiple straight lines that link breakpoints of geneA and geneB transcripts (plot downloading version)}
\usage{
plot_separate_domain_download(
  first,
  first_name,
  first_domain,
  first_motif,
  second,
  second_name,
  second_domain,
  second_motif,
  breakpoint
)
}
\description{
Plot transcript-model(reducing introns), domain, motif and multiple straight lines that link breakpoints of geneA and geneB transcripts together.
}
