% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_persample.R
\name{plot_separate_individual}
\alias{plot_separate_individual}
\title{Plot transcript-model and chromosome ideogram of partner genes}
\usage{
plot_separate_individual(
  first,
  first_name,
  second,
  second_name,
  breakpoint,
  cytoband
)
}
\arguments{
\item{first}{A list: \strong{'key'} is a string \code{'\$pos'} (i.e. breakpoint pos in geneA); \code{'first\$pos\$transcript'} is a data.frame object
(only one row) that collects evaluation of breakpoint pos in geneA transcript; \code{'first\$pos\$select_region'} is a data.frame object
(5utr-cds-utr3 annotation) for plotting and it is a constant variable for different \code{'\$pos'} values.}

\item{first_name}{A string - symbol name of geneA (e.g. 'TMPRSS2').}

\item{second}{A list: \strong{'key'} is a string \code{'\$pos'} (i.e. breakpoint pos in geneB); \code{'second\$pos\$transcript'} is a data.frame object
(only one row) that collects evaluation of breakpoint pos in geneB transcript; \code{'second\$pos\$select_region'} is a data.frame object
(5utr-cds-utr3 annotation) for plotting and it is a constant variable for different \code{'\$pos'} values.}

\item{second_name}{A string - symbol name of geneB (e.g. 'ERG').}

\item{breakpoint}{A data.frame with three columns (e.g. \code{'breakpoint[i,1]'} - breakpoint pos in geneA; \code{'breakpoint[i,2]'} - breakpoint pos in geneB;
\code{'breakpoint[i,3]'} and \code{'breakpoint[i,4]'} are the number of split and span reads supporting this fusion).}

\item{cytoband}{A GRange object for plotting cytoband of chromosome ideogram.}
}
\description{
Plot transcript-model and chromosome ideogram of partner genes in 'per_sample' tab-panel (for downloading version as well)
}
