% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_persample_bam.R
\name{plot_separate_individual_bam}
\alias{plot_separate_individual_bam}
\title{Plot fusion transcript in context of chromosome ideogram, gene annotation and read coverage}
\usage{
plot_separate_individual_bam(
  first_name,
  second_name,
  breakpoint_A,
  breakpoint_B,
  rna_bam_path,
  dna_bam_path = NULL,
  coverage_plot_trans = F,
  offset = 5,
  chrom_notation_rna = T,
  chrom_notation_dna = T,
  split = NULL,
  span = NULL,
  transcriptA = NULL,
  transcriptB = NULL,
  fusion_strandA = NULL,
  fusion_strandB = NULL,
  coverage_max_A = NULL,
  coverage_max_B = NULL
)
}
\arguments{
\item{first_name}{A string - symbol name of geneA (e.g. 'TMPRSS2').}

\item{second_name}{A string - symbol name of geneB (e.g. 'ERG').}

\item{breakpoint_A}{A number - genomic coordinate of breakpoint in geneA.}

\item{breakpoint_B}{A number - genomic coordinate of breakpoint in geneB.}

\item{rna_bam_path}{A string - path of alignment file from RNA-seq (BAM format).}

\item{dna_bam_path}{A string - path of alignment file from DNA-seq (BAM format), default value: NULL.}

\item{coverage_plot_trans}{A boolean type - if TRUE, RNA read coverage calculated by reads mapped to selected transcripts of geneA and geneB; otherwise (default: FALSE) RNA read coverage calculated by reads mapped to genomic regions of geneA and geneB. coverage_plot_trans is only valid for RNA-seq alignment.}

\item{offset}{A number - the value (default: 5) indicates how many bases are extended from 5'-end and 3'-end of geneA and geneB in plotting.}

\item{chrom_notation_rna}{A boolean type - if TRUE (default value) chromosome name in RNA-seq alignment denotes like 'chrX'; if FALSE chromosome denotes like 'X'.}

\item{chrom_notation_dna}{A boolean type - if TRUE (default value) chromosome name in DNA-seq alignment denotes like 'chrX'; if FALSE chromosome denotes like 'X'.}

\item{split}{A number - split reads support (default: NULL).}

\item{span}{A number - discordant spanning read pair support (default: NULL).}

\item{transcriptA}{A vector of string - selected transcripts (ensembl id) of geneA for plotting (default: NULL).}

\item{transcriptB}{A vector of string - selected transcripts (ensembl id) of geneB for plotting (default: NULL).}

\item{fusion_strandA}{A string - the strand direction of geneA in transcribed fusion (default: NULL; options: '+' or '-').}

\item{fusion_strandB}{A string - the strand direction of geneB in transcribed fusion (default: NULL; options: '+' or '-').}

\item{coverage_max_A}{A number - the range of y-axis (RNA-seq read coverage) scale for geneA (default: NULL).}

\item{coverage_max_B}{A number - the range of y-axis (RNA-seq read coverage) scale for geneB (default: NULL).}
}
\description{
Plot curved lines to link breakpoints of fusion transcripts in context of gene annotation and read coverage
}
