% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_overview.R
\name{plot_separate_overview}
\alias{plot_separate_overview}
\title{Plot transcript-model and chromosome ideogram of partner genes}
\usage{
plot_separate_overview(first, first_name, second, second_name, cytoband)
}
\arguments{
\item{first}{A list: \strong{'key'} is a string \code{'\$pos'} (i.e. breakpoint pos in geneA); \code{'first\$pos\$transcript'} is a data.frame object
(only one row) that collects evaluation of breakpoint pos in geneA transcript; \code{'first\$pos\$select_region'} is a data.frame object
(5utr-cds-utr3 annotation) for plotting and it is a constant variable for different \code{'\$pos'} values.}

\item{first_name}{A string - symbol name of geneA (e.g. 'TMPRSS2').}

\item{second}{A list: \strong{'key'} is a string \code{'\$pos'} (i.e. breakpoint pos in geneB); \code{'second\$pos\$transcript'} is a data.frame object
(only one row) that collects evaluation of breakpoint pos in geneB transcript; \code{'second\$pos\$select_region'} is a data.frame object
(5utr-cds-utr3 annotation) for plotting and it is a constant variable for different \code{'\$pos'} values.}

\item{second_name}{A string - symbol name of geneB (e.g. 'ERG').}

\item{cytoband}{A GRange object for plotting cytoband of chromosome ideogram.}
}
\value{
A list with two data.frame elements (i.e. \code{'A1_xy'} and \code{'B1_xy'}) that contain exon coordinates in the plot layout.
}
\description{
Plot transcript-model and chromosome ideogram of partner genes in 'overview' tab-panel
}
